Squashed commit of the following:
commit a4f4d36b0a40e0854b088b5ea1048ff2dfc35081
Author: JustSergioPB <justsergioit.com>
Date:   Fri Oct 11 11:39:19 2024 +0200

    feat(NOT-2): Hardlock for actions

commit 9d4bb6994ddb3df1f2080aafdb4b75cf6bdb3c10
Author: JustSergioPB <justsergioit.com>
Date:   Thu Oct 10 18:32:45 2024 +0200

    feat(NOT-2): Filtering by org

commit e68e81d062dc7c6f868187ce97f4f073f27b4bdb
Author: JustSergioPB <justsergioit.com>
Date:   Thu Oct 10 18:16:13 2024 +0200

    feat(NOT-2): Stopped state

commit 46d158ccb21a3baddaaf01ac8108f5c210107ca0
Author: JustSergioPB <justsergioit.com>
Date:   Thu Oct 10 17:51:03 2024 +0200

    feat(NOT-2): Soft lock for non active users

commit 1828a85328188da4a5d9057d21a11338072ec50b
Author: JustSergioPB <justsergioit.com>
Date:   Thu Oct 10 16:49:21 2024 +0200

    feat(NOT-2): Org admin created by default

commit 30865aeda0c88e0a81acd37e240d371cc249176d
Author: JustSergioPB <justsergioit.com>
Date:   Thu Oct 10 16:39:05 2024 +0200

    feat(NOT-2): Confirmation cycle implemented

commit 23ecd20fa68261c21987702bd437bf3e01898915
Author: JustSergioPB <justsergioit.com>
Date:   Thu Oct 10 14:20:02 2024 +0200

    feat(NOT-2): User translations + email verification

commit cf1502a0a1b409d6dc62b51f09ddba84bb93fcb0
Author: JustSergioPB <justsergioit.com>
Date:   Thu Oct 10 12:12:58 2024 +0200

    feat(NOT-2): Working registration

commit 4fe31d0a331b30332c82568486721f8c2c10cdda
Author: JustSergioPB <justsergioit.com>
Date:   Tue Oct 8 11:28:18 2024 +0200

    feat(NOT-2): Created form and refactored context

+++ lib/notebook_server_web/live/org_live/index.ex
@@ -3,6 +3,7 @@ defmodule NotebookServerWeb.OrgLive.Index do
 
   alias NotebookServer.Orgs
   alias NotebookServer.Orgs.Org
+  alias NotebookServer.Accounts.User
   use Gettext, backend: NotebookServerWeb.Gettext
 
   @impl true
@@ -40,25 +41,44 @@ defmodule NotebookServerWeb.OrgLive.Index do
 
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
-    org = Orgs.get_org!(id)
-    {:ok, _} = Orgs.delete_org(org)
-
-    {:noreply, stream_delete(socket, :orgs, org)}
+    if User.can_use_platform?(socket.assigns.current_user) do
+      org = Orgs.get_org!(id)
+      {:ok, _} = Orgs.delete_org(org)
+      {:noreply, stream_delete(socket, :orgs, org)}
+    else
+      {:noreply, socket}
+    end
   end
 
   @impl true
   def handle_event("deactivate", %{"id" => id}, socket) do
-    org = Orgs.get_org!(id)
-    {:ok, _} = Orgs.deactivate_org(org)
-
-    {:noreply, stream(socket, :orgs, Orgs.list_orgs())}
+    if User.can_use_platform?(socket.assigns.current_user) do
+      org = Orgs.get_org!(id)
+      {:ok, _} = Orgs.deactivate_org(org)
+      {:noreply, stream(socket, :orgs, Orgs.list_orgs())}
+    else
+      {:noreply, socket}
+    end
   end
 
   @impl true
   def handle_event("activate", %{"id" => id}, socket) do
-    org = Orgs.get_org!(id)
-    {:ok, _} = Orgs.activate_org(org)
+    if User.can_use_platform?(socket.assigns.current_user) do
+      org = Orgs.get_org!(id)
+      {:ok, _} = Orgs.deactivate_org(org)
+      {:noreply, stream(socket, :orgs, Orgs.list_orgs())}
+    else
+      {:noreply, socket}
+    end
+  end
 
-    {:noreply, stream(socket, :orgs, Orgs.list_orgs())}
+  def handle_event("stop", %{"id" => id}, socket) do
+    if User.can_use_platform?(socket.assigns.current_user) do
+      org = Orgs.get_org!(id)
+      {:ok, _} = Orgs.stop_org(org)
+      {:noreply, stream(socket, :orgs, Orgs.list_orgs())}
+    else
+      {:noreply, socket}
+    end
   end
 end
